remove glGetError wrapper
We were trying to hide any internally generated errors from the
application in case they might cause it to quit for example, but
the wrapper wasn't really able to do this correctly. The problem
is that we really need to have a callback from GL whenever
errors are generated (e.g. via KHR_debug) so that when we know
when to throw away any internally saved error values.

For now instead we will just try to never generate internal errors and
only look at a more elaborate solution via KHR_debug if that approach
doesn't work.