perf: codegen OA counter normalization code
This replaces the hand written counter normalization and query building
code with code generated via a python script from an XML description of
the metric sets and counters.

The XML file is the same as I've also added to my wip/rib/oa-hsw-codegen
branch of Mesa.

This should make it very easy to also experiment with further sets of
metrics, such as compute metrics.