Add support for experimental i915 perf OA interface
In parallel with the existing interface based on Linux perf this adds
support for an alternative i915 specific event interface for accessing
OA metrics.

Supporting both in gputop makes it easier us to compare the merits of
each approach and avoids the need to switch branches of gputop to
test the different drivers.

gputop will first try and open OA events via perf_event_open but if that
fails it will try the i915_perf interface instead.