update to latest i915 perf API + rm core perf OA support
The latest i915 perf API takes an array of uint64 properties[] when
opening a stream which this adds support for.

Now that we've decided to aim for a drm/i915 specific interface for
exposing OA metrics this drops support for the previous core perf based
OA driver. Some core perf related code has been left behind assuming
we will still want to use it to access cpu metrics.