build: avoid cyclical libtool dependency
gputop/Makefile.am was adding -lGL to libgputop_la_LIBADD, intending to
link against the *system* libGL, but because the same Makefile.am is
also responsible for building the gputop wrapper libGL libtool
interprets that it should link with that instead. Since the wrapper
depends on libgputop.so that then resulted in a cyclical dependency.

This ensures libgputop.so can be built without being explicitly linked
with libGL at build time and it instead it will open the system libGL.so
using dlopen.