gl: enumerate + add ui tabs for all GL perf queries
This also changes the interaction between the GL and ncurses UI thread
to be in terms of a linked list of 'finished' queries prepared by the GL
thread for the UI thread, and a cache of query objects which the UI
thread returns objects to once it has visualized them.

UI-wise this notably isn't very pleasant a.t.m. We now have quite a
large number of tabs that can't all be seen and it's awkward to have to
cycle through them sequentially to move to the tab you want.

For testing purposes though this at least give a way to smoke test more
than just the Render Basic GL performance query.

Later it would be better to support visualizing GL queries with the Web
UI.