Implement context interception and tracking
Here we intercept the drm ioctl' requests to create/destroy
contexts. As new requests come in we record the fd + ctx_id pair
and add them to a global list of contexts. We also manage the
destruction of contexts here.

This work will later enable us to perform per-context
monitoring without the gl intercept work.