Consider context-switch on GEN8+
On GEN8+ the hardware generates a report to indicate when a context-switch
occurs. This should now be handled properly, as we now use this special report
as the base for subsequent accumulation calculations.

V2: set the gen for the different cases we handle in init_dev_info(), to
minimize the separate utilities that may need updating for new HW
enabling(rib)