gl: implement dlopen() LD_PRELOAD wrapper
To account for applications/frameworks that dlopen("libGL.so",) at
runtime (such as SDL) we wrap dlopen and instead return a handle to
dlopen(NULL,) which is effectively a handle to our libfakeGL.so that's
forcibly loaded via LD_PRELOAD.

Alternatively we could consider going back to generating a full
libGL.so.1 wrapper library at some point, except without depending on
the ifunc mechanism.

Note for SDL we sill need to follow up with a few other fixes, such
as ensuring a glXQueryExtension symbol can be found in libfakeGL.so
and improving our glXGetProcAddress wrapper to recognise more than
just the CreateContext functions need wrapping.