rework how streams and their libuv handles are closed
The previous code mistakenly thought that uv_poll_stop or uv_timer_stop
were enough to disconnect a uv_poll_t or uv_timer_t from libuv before
being freed but actually it's necessary to call uv_close() to both stop
the handle and unlink it from the mainloop.

The fiddly thing here is that uv_close() doesn't complete synchronously,
it takes a callback which is called at some later time and we need to
make sure to delay freeing stream resources until all of it's handles
have closed via the mainloop.

This adds a gputop_perf_stream_close() api to explicitly close a stream
instead of doing this implicitly when the last reference is dropped.
It's now a requirement to have closed a stream before dropping the last
reference. Similar to the uv_close() api, gputop_perf_stream_close()
also takes a callback to notify completion.

gputop-server.c and gputop-ui.c have then also been updated to use this
new api.