gputop-ui: fix current_tab reset flow
When closing a tab we should always reset the current_tab, regardless of
the value of gputop_current_perf_stream, otherwise we can get stuck with
the ui not drawing the pending_tab for GL queries, since GL queries
don't affect the value of gputop_current_perf_stream. This is only
applicable when running with perf.paranoid enabled.