rework per-ctx mechanism
Instead of linking the per_ctx_mode with a query, it makes more sense to
use the stream to store this mode. A gputop_perf_query object should be
considered immutable. As a result we now also pass the ctx_handle
directly to the gputop_open_i915_perf_oa_query, a NULL ctx here implies
that a system-wide stream will be opened. Awkwardly the web version has zero
knowledge of the stream object, so we end up piping the per_ctx_mode
through the gputop_worker_query, in order to properly handle reports
generated because of a CTX_SWITCH in the web-worker.

Some of the error handling code has also been cleaned up with this patch.