gputop-ui: Adding support for PID and Context ID display
Adding changes into the CSS to deal with better margins.

On this commit we introduce the Demo version

We define the domains in demo mode supported by default in the index.html

There is also a ?demo=1 parameter that you can append to the index
so you can force the demo mode.

Flags at this moment break the demo mode functionality

The pid is only being supported by the UI and the communication
is working. The pid internal functionality will be added after the kernel
supports this functionality.