devinfo: initialize timestamp_frequency
It used to be that we could assume the least significant bit of the gpu
timestamps equated to 80ns, but for SKL/BSW+ things have changed and
it's also no longer a neat integer multiple. We're now going to start
normalizing raw gpu timestamps based on a timestamp_frequency constant
instead of scaling constant.