remove gettext dep
We aren't using it with the ncurses ui so far and wouldn't expect to use
it with the web ui. For now it's just a pain with distros having
different gettext versions and have to to re gettextize to be able to
build, but touching files that have to be tracked in git.