Gputop.js: add rpc_request api
This generalises, setting msg.uuid, encoding and sending an RPC message,
with some debug logging and optionally associating a closure with a
request to be called once we receive a reply.

This switches to using msg.set(key, value) for setting protobuf.js
message members to ensure the proper handling of 'oneof' properties
which also gives us a something to print for debugging purposes.