Scale duration dependent metrics into per-second values
Normalized metrics are calculated based on deltas that have been
accumulated over some known period of time, but that period may not be
directly meaningful and currently might vary depending on the zoom level
of the tracing graphs. To give a concrete frame of reference the metrics
are now scaled with respect to a one second duration and given a '/s'
per-second suffix as appropriate.

Later it might be useful to be able to specify a reference duration, so
e.g. if you are targeting a certain fps it might be useful to see other
duration based metrics with respect to your per-frame budget.

It should be noted that we're currently compounding the error margin of
the values we're presenting as we're typically scaling up metrics that
have only been accumulated over very short durations. The plan to
address this is that we should have separate accumulation state for
calculating values we can plot on the graphs vs the per-second text
values we show.