Commit gziped gputop-web.js[.mem] to repo
As a practical compromise to lower the bar for others building and
testing the gputop web ui without needing to setup the Emscripten
toolchain (which involves a slow build of LLVM) this adds a compressed,
pre-built gputop-web.js.

configure.ac now sets a HAVE_EMCC automake conditional which affects
whether gputop-web.js is derived by decompressing gputop-web.js.gz or
calls through to Makefile.emscripten

The compressed gputop-web.js.gz is ~140k so hopefully we won't bloat the
repo size to quickly with this. Extra care is taken to check the md5sum
of the input source with any existing compressed source to avoid
redundant gputop-web.js.gz updates.

For now all gputop-web.js builds compile with -Oz to aggressively
optimize for size and without -g3/4 so no source maps are generated.
Later we might want to make these flags controllable for development.