Add http_get() utility for http request
This adds a common utility api for making http GET requests that works
consistently with node.js and in a browser.

One of the notable side effects of this change is to avoid explicitly
appending a port number to the url used to request gputop-web.js which
doesn't work in demo mode.