remove --nodejs cmd line option for now
Instead of supporting launching ncurses tools via the gputop binary the
gputop-csv tool can currently be run directly.

The main limitation with this currently is that gputop-csv can only connect to
an already running workload/server but for this kind of command line
tool that's probably acceptable for now.

Supporting launching node.js tools via the gputop binary is currently a
little fiddly/fragile from the pov of ensuring the server is reliably
shutdown if the client exits.