Allow building gputop without the web/node.js clients
Currently it's a bit of a pain to set up Emscripten for building the
web/node.js client code (depends on fastcomp llvm branch). To lower the
barrier for others trying out Gputop this allows building without the
web/node.js clients but still supporting remotely accessing metrics via
a websocket - which should be enough to use the UI hosted on
gputop.github.io.