update how system properties are forwarded to webc code
Instead of having a function where we have to keep extending the
arguments passed, we have a function to reset the system properties
and then a function for setting individual properties.

protobuf reflection is now used to automatically forward all
Features.DevInfo properties.

_Generic is used in the webc code to help double check the range of
values before setting a property, in case a struct gputop_devinfo
member type is changed and we overlook something here.

Finally this is intended to make it simpler to bind the webc code in a
node.js addon, so we don't also have to modify the bindings to add new
properties.