Add scripts for maintain OA XML files + codegen
Maintaining these as part of gputop seems as good a place as any for
now.

mdapi-xml-convert.py:
---------------------
 "MDAPI" xml files are an XML schema for maintaining meta data about Gen
 graphics Ovservability counters, where MD API is the name of a library
 shared by Intel GPA and Intel VTune.

 These files aren't publicly documented and have some historical baggage
 that adds some complexity as well as being inconsistent in a number of
 ways that makes it quite a bit of effort to parse/use the data. We also
 don't have guarantees about how this schema is maintained.

 This script takes the opportunity to find ways to simplify the input
 data and to make it more consistent to hopefully reduce the effort
 involved in using the data downstream. This generates the oa-*.xml
 files that gputop, drm i915-perf and Mesa use.

 If mdapi xml files are copied under gputop-data/mdapi/* then the
 gputop build will run this script to update oa-*.xml files.

i915-perf-kernelgen.py:
-----------------------
 Based on the oa-*.xml files this generates code for metric sets
 supported by the drm i915-perf driver including:

 - Static arrays describing the various NOA/Boolean/OA register configs
 - Functions/structs for advertising metrics via sysfs
 - Code that can evaluate which configs are available for the current system
   based on the RPN availability equations

update-guids.py:
----------------
 Helps maintain the registry of hardware config GUIDS in guids.xml (also
 added in this commit):

 - Automatically adds template entries for unregistered metric sets discovered
   in new mdapi xml files.
 - Once mdapi-convert-xml.py has been run to output register configs for new
   metric sets then re-running this script can add the 'v2' config hash to
   corresponding registry entries.

This commit brings with it updated oa-*.xml files based on these
scripts, with some minor corresponding changes considering a rename of
the metric set 'guid' attribute to 'hw_config_guid' to clarify what the
ID is really for. The counter for the average gpu frequency was also
renamed to not include units in the name.

This also adds a new oa-bxt.xml file in preparation for Broxton support.

The rules for running these scripts are in gputop-data/Makefile.xml to
ease running them in isolation, e.g. to generate the files for i915-perf
while not working with gputop itself. Makefile.xml supports some codegen
options like SYSFS=1 e.g.:

$  make -C gputop-data -f Makefile.xml SYSFS=0

Check Makefile.xml for details