update to expect 0xffffffff as invalid ctx id in switch-away reports
When the kernel is filtering OA reports for a specific context then
whenever the HW switches context the kernel forwards the first report
after each switch (tagged by the HW with the next context ID) but with
the real ID scrubbed out.

These reports used to have an ID of 0x1fffff, but that was poorly chosen
since execlist IDs are 21 bits not 20 bits making this a valid HW ID.
Bit 21 is currently reserved for future use whereby 0x3fffff may not be
wise either. The kernel now writes 0xffffffff as an invalid ID.