configure.ac updates
Be a bit more conservative about enabling building the web ui or node
tools so a successful build is more likely.

The web ui is disabled by default since it's a bit of a faff to set up
Emscripen on Linux, including building llvm-fastcomp. We don't
automatically enable based on whether emcc is found since we know that
the Emscripten package on Ubuntu isn't usable (maybe the most likely way
we'd find emcc, unless someone was really intent on building the gputop
web ui)

The node tools are automatically enabled based on whether node >= v6,
node-gyp >= v3 and npm are all found.

The message at the end of configuring is tailored to show using the
web UI hosted on github depending on whether the web UI is enabled.