kernelgen: avoid advertising a metric that the current hw sku can't support
Previously we advertised all metric sets we knew about for a particular
hardware generation but only evaluated whether the current hardware sku
supported the metric set when the user attempted to open it. Now we
avoid advertising the metrics in the first place if we know the current
system does't support it.