Support multiple OA counter accumulators
The 'accumulator' in gputop is a state machine that is fed raw OA
reports and calculates deltas between sequential reports to build up an
array of uint64_t counter values over an accumulation period before
this array is fed as input into counter normalization functions.

To be able to accumulate values over a long period for readable labels
in a UI (e.g. one second) and at the same time accumulating over a very
short period for a detailed graph plot this enables us to maintain
multiple accumulators that all raw reports are fed into.

The intention is that these accumulators will later be capable of
more advanced context filtering, but for now the previously attempted
support for single-context filtering has been removed.