gputop-csv: a number of UI improvements
By default treat it as an error to give an unknown counter/column name,
unless -u/--allow-unknown-columns given (potentially useful for
automated profiling systems).

If an unknown counter is given then gputop-csv will automatically list
what counters are available for the given metric set.

Counters are listed alphanumerically now, with some padding to hopefully
be a bit easier to scan.

The default output is less verbose now, though --debug can show all the
internal logging.

The stderr header showing the configuration now just prints the
string given to --columns instead of the human readable names with
quotes so it's easy to copy & paste for someone else reproducing the
same configuration.