gputop-client-c: serialize oa-*.[ch] codegen
We daisy chain these oa-*.[ch] dependencies so only the header directly
depends on the script to be generated and the .c file depends on the
header instead of using a multi target rule.  This ensures that a
parallel make won't resort to running the script twice (with the risk
that the two runs trample on each other generating the same files).