gputop-client-c: more parallel make fixes
This previous approach of daisy chaining the oa-*.c and oa-*.h target
deps was that the oa-*.c target didn't have it's own rule so if just
the .c file was missing it wouldn't be regenerated. This gives each
file a dedicated rule instead.

This also ensures we don't call though to Makefile.emscripten to build
gputop-web.js until all the oa-*.[ch] files are generated.