update how we choose an OA sample period for an accumulation period
This implements a more sophisticated function for choosing an OA
sampling period for a requested accumulation period within a given
margin of error and also the ability to constrain the maximum OA
sampling period.

This is for dealing with the limited exponential scale for programming
the OA unit sampling period while giving general freedom to pick any
accumulation period. For an accumulation period that the hardware
doesn't support and if the closest that the hardware supports is
significantly different then this will instead try and divide the
accumulation period into units of a smaller OA sampling period.

The constraint can be used to disallow long hardware sampling periods
where 32bit counters might overflow multiple times within that time.

Note: the gputop-csv command line UI has been changed in a backwards
incompatible way so that the -p/--period arguments now specify an
accumulation period. Overriding the OA unit sampling period is now
clearly discouraged but there is still a --oa-sample-exponent for
being able to force a hardware sampling period if needs be.