gputop-csv: condense cols, print sys info, re-print titles
To help minimize the horizontal space that gputop-csv results take then
if the units for a counter are "percent" we simply print "%" and now
each column may have a different width.

To hopefully help ensure we're comparing apples with apples when sharing
gputop-csv results around we not output more details about the system
that was used to capture the data.

If we're paging through lots of data in a terminal then we now re-print
the titles based on the current terminal height to each reading.