server: drop usage of finish_callback
In :

commit f04108a832cfa765ae218e15c407f7a80600ae0a
Author: Robert Bragg <robert@sixbynine.org>
Date:   Tue Jul 21 14:51:44 2015 +0100

    wslay: add 'finish' callback for _event_omsg for app to free source data

Wslay was modified to add a callback notifying that a msg wasn't used
anymore (usually it's been sent). This drop the usage of this callback
by doing its job in the read_callback when it reaches the EOF.

This allows us to remain on a much closer to upstream version of
wslay.