meson: Add peg subproject
Peg is a somewhat obscure tool that's not reliably packaged across
distributions (for example Arch Linux only has an AUR package for
the 0.1.15 release from 2013).

This adds a subproject for building the leg tool used to generate
tracepoint_format.leg.h so users don't have to figure out how to
install the correct version.