meson: rework UI selection
On my distro (Arch), GLFW only supports one platform at a time, so you
can either have X11 or Wayland.

Meson previously detected the presence of x11 to decide whether to build
the x11 UI, which wouldn't compile on my system running wayland with the
wayland GLFW, even though libX11 was installed.

A better way of handling this is to let the user decide which platform
to build, which this commit does. The option is an array, which means
multiple UIs can be selected at the same time, and the build's success
will depend on what the user's system allows.

Signed-off-by: Eric Engestrom <eric.engestrom@intel.com>