data: add query only counters
Some metric counters rely on snapshot of registers that are not copied
into the OA reports such as PERF_COUNTER_[12].

Those will not be available in gputop but they could be in Mesa.

To deal with those counters we flag them with the availability on the
special variable that tells use when we're in query mode.